refactor(find_contacts): #557 — extract scripts/find_contacts.py to findajob.find_contacts (M3+)
Same shape as M3's four script→package extractions (#539/#541/#542/#543).
Logic moves to src/findajob/find_contacts.py (single module — 245 LOC,
well under the 300 LOC soft cap). scripts/find_contacts.py shrinks to a
10-LOC entry-point shim. Subprocess invocation from prep/orchestrator
unchanged.

Module-load load_env() moved into main() per the M3 import-safety
discipline. New tests/test_find_contacts_import_safety.py characterizes:
load_env not called at import, public surface (main + logic helpers)
exposed, company_match() blank-string guard preserved.

Existing tests (test_find_contacts.py #184/#164, test_find_contacts_cost_logging.py)
import-paths updated. Type annotations on the relocated module's public
surface — strict mypy satisfied under M7's gate.

First M3+ extraction PR (the four prior were M3 proper).

Verification:
- uv run pytest -q → 1975 passed, 1 skipped (pre-existing)
- uv run ruff check src/ scripts/ tests/ → All checks passed
- uv run ruff format --check → clean
- uv run mypy src/findajob/ → no issues found in 114 source files

Closes #557.
