+++ CHANGELOG.md
@@ -11,6 +11,8 @@ changes may land in minor version bumps; patch releases are bugfix-only.
 ## [Unreleased]
 
 ### Refactor
+- **#557 refactor(find_contacts): extract `scripts/find_contacts.py` to `findajob.find_contacts` module (M3+).** Same shape as the four script→package extractions M3 already shipped (#539 / #541 / #542 / #543). Logic moved to `src/findajob/find_contacts.py` (single module — file is 245 LOC, well under the 300 LOC soft cap, no need to split further). `scripts/find_contacts.py` shrinks to a 10-LOC entry-point shim that imports `main` from the package; subprocess invocation from `findajob.prep.orchestrator` is unchanged. Module-load `load_env()` (line 47 of the original) moved into `main()`, locking the import-safety property characterized by the new `tests/test_find_contacts_import_safety.py` (4 tests: load_env not called at import, `main`/logic-helpers exposed, blank-string guard preserved). Existing `tests/test_find_contacts.py` (#184 / #164 regressions) and `tests/test_find_contacts_cost_logging.py` (Phase 2 cost-log tests) updated to import from the new path. Type annotations added on the public surface — strict mypy is satisfied for the relocated module under M7's gate. First M3+ extraction PR.
+
 - **#556 refactor(prep_naming): consolidate `abbrev_title()` duplication (M3+).** First M3+ punch-list item from the closed M3 epic (#537). `abbrev_title()` was duplicated between `findajob.prep.orchestrator` (typed) and `scripts/rename_folders.py` (untyped); both copies kept in sync by code review rather than code structure. Consolidated to `findajob.prep_naming` (the natural home — same domain as `safe_filename_part`, both build folder-safe filename components). Both call sites import from there. `tests/test_prep_import_safety.py::test_orchestrator_exposes_main_and_abbrev_title` renamed to `test_orchestrator_exposes_main` (orchestrator's deliberate public symbol is `main`; `abbrev_title` is now an import); `test_abbrev_title_behavior` updated to import from the canonical location, mirroring the no-re-export-theater discipline established by the M3 cleanup PR (#544).
 
 ## [0.21.0] — 2026-05-08
